requirements.txt
crawl4ai
fastapi
fastui
uvicorn
logfire[fastapi]
mongoengine
pydantic_settings